feat: render searched movies
