+++ src/components/Movies/MoviesCard/MoviesCard.js
@@ -1,13 +1,13 @@
 import React from 'react';
 import { Route } from 'react-router';
-import { CurrentUserContext } from '../../../contexts/CurrentUserContext';
+// import { CurrentUserContext } from '../../../contexts/CurrentUserContext';
 
 function MoviesCard(props) {
 
-  const currentUser = React.useContext(CurrentUserContext);
+  // const currentUser = React.useContext(CurrentUserContext);
 
-  const isSaved = props.movie.owner === currentUser._id;
-  const movieCardClassName = `button button_type_save ${isSaved ? 'button_type_save-active' : ''}`;
+  // const isSaved = props.movie.owner === currentUser._id;
+  const movieCardClassName = `button button_type_save ${props.isSaved ? 'button_type_save-active' : ''}`;
 
   function handleSaveClick() {
     props.onSaveClick(props.movie);
